transcendence_src/requirements.txt
Django
psycopg2-binary
djangorestframework
markdown
django-filter